created widget files for sample summary and texture button
